renaming files to match class name
