+++ src/monitor/AgentsExplorer.cpp
@@ -1,7 +1,7 @@
 
 #include <QDebug>
 
-#include "AgentsEnumerator.hpp"
+#include "AgentsExplorer.hpp"
 #include "common/constants.hpp"
